requirements.txt
pandas
seaborn
matplotlib
sqlalchemy
pymysql  # Librería para conectarse a MySQL
python-dotenv  # Para manejar variables de entorno desde el archivo .env
requests